+++ customize/menu.sh
@@ -18,12 +18,12 @@ customize_menu() {
             "Back")
 
         case "$choice" in
-            1) source "$MACRIFT_DIR/apps/terminal.sh" && terminal_menu ;;
-            2) source "$MACRIFT_DIR/apps/terminal.sh" && shell_menu ;;
-            3) source "$MACRIFT_DIR/apps/terminal.sh" && fastfetch_menu ;;
-            4) source "$MACRIFT_DIR/apps/editor.sh" && editor_menu ;;
-            5) source "$MACRIFT_DIR/apps/dock_layout.sh" && dock_layout_menu ;;
-            6) source "$MACRIFT_DIR/apps/wallpaper.sh" && wallpaper_menu ;;
+            1) source "$MACRIFT_DIR/customize/terminal.sh" && terminal_menu ;;
+            2) source "$MACRIFT_DIR/customize/terminal.sh" && shell_menu ;;
+            3) source "$MACRIFT_DIR/customize/terminal.sh" && fastfetch_menu ;;
+            4) source "$MACRIFT_DIR/customize/editor.sh" && editor_menu ;;
+            5) source "$MACRIFT_DIR/customize/dock_layout.sh" && dock_layout_menu ;;
+            6) source "$MACRIFT_DIR/customize/wallpaper.sh" && wallpaper_menu ;;
             0) break ;;
             *) ;;
         esac