+++ tweaks/input.sh
@@ -12,6 +12,7 @@ input_tweaks() {
 
     audit_default "com.apple.AppleMultitouchTrackpad" "TrackpadRightClick" "-bool" "true" "Right-click"
     audit_default "NSGlobalDomain" "NSWindowShouldDragOnGesture" "-bool" "true" "Drag windows anywhere"
+    audit_default "NSGlobalDomain" "com.apple.mouse.linear" "-bool" "true" "Disable pointer acceleration"
 
     [[ "$MACRIFT_BATCH_TWEAKS" == true ]] && return 0